add PublicKeyExt trait
