+++ crates/iota-sdk/examples/generate_ed25519_address.rs
@@ -3,6 +3,7 @@
 
 use base64ct::{Base64, Encoding};
 use iota_crypto::{PrivateKeyExt, ed25519::Ed25519PrivateKey};
+use iota_types::PublicKeyExt;
 use rand::rngs::OsRng;
 
 fn main() {
